// frontend/src/app/pages/model/infoBloqW.ts
export class InfoBloqW {
  id: number;
  week: number;
  year: number;
  fmq: string;
  sujetSupport: string;
  qui: string;
  bloquantEuro: string;
  actMenees: string;
  delai: string; // [s1,s2] ou RAS
  tauxAbst: string;
  escalader: string;
  abstRaison: string;
  atelier: string;

  constructor() {}
}